Refactor force_sub method and add randomization options
The developer refactored the `force_sub` method in the `UserView` class and added two new parameters: `username` and `randomize`. The method now checks if the `username` and `randomize` parameters are provided in the request arguments and uses them accordingly. Additionally, the developer made changes to the `get_proper_config` method to also accept the `username` and `randomize` parameters. The `all_configs` method was modified to handle the new `randomize` parameter, allowing for randomization of server or config options. The changes also include retrieving additional configurations from external sources specified in a `nodes.json` file.